feat: measure selection is out of the layer selector
